Use Symfony's ProcessBuilder to safely build phpunit command.
Spaces in the arguments (including in the filepath to the paratest
executable) were causing paratest to fail.  This uses the ProcessBuilder
to build the properly escaped command to execute.

I would like to use the process returned by the processbuilder
throughout the ExecutableTest class, but that would involve breaking the
public API of this class.  I left that as a todo for a major version
change and for the owners of the project to decide on.

This change shouldn't break things, but I don't know Windows well enough
to know if it would break due to the command being in quotes.