Fix version reporting in console, closes #257
ParaTest has the version number hard-encoded.

Previously when reporting the version via --version that hard-encoded
version number was presented in the output (e.g. 1.0.1).

As technically the version changes with git commit and not every git
commit is tagged and not even every tagged commit contains the appropriate
change to the hard-encoded version number within

    \ParaTest\Console\ParaTestApplication::VERSION

the wrong version is reported.

This commit addresses the issue by first looking up the installed version
of the package "brianium/paratest" from within composer's installed.json
file which contains the currently installed version (based on the git tag,
see Packagist) and if not found (e.g. you have checked out the development
version in a git clone) the version number will be taken from git directly
(e.g. 1.0.1-10-g89f72e2, see man git-describe).

If all of these look-ups fail, the fallback is the current behavior of
taking the version from the constant.

Refs:

- #257