Fixed problem composer installs with bin-compat full
This will fix the method for determining the phpunit path, when installing e.g. on windows with the composer configuration "bin-compat": "full" and running tests inside a linux docker.  
The problem was that the current sollution determines the phpunit executable file based on the existence of the ".bat"-file. This solution will determine the executable based on file existence and the systems DIRECTORY_SEPERATOR.