#310 Implement new SqliteRunner which uses sqlite to dispatch tests.
This commit implements a SqliteRunner which attempts to resolve the
issues faced in #310 with the WrapperRunner. It's derived from the
WrapperRunner and dispatches n workers upon start, which then
independently fetch pending tests from a pre-populated sqlite database
until all tests are completed.

This commit also changes the way the runners are chosen during
bootstrapping. It will now dynamically check for the class in the
default namespace.

In addition, this commit will remove the isRunning flag from the
WrapperRunner because accessing the process information is more
reliable.

closes #310