Update WrapperRunner.php
When thare are no tests to run, the `LogMerger` crashes on the last test (`assert($mainSuite !== null);`), because there are no log files generated to merge together. With this initial check in the WrapperRunner, we only try to merge the log files if there are actually created log files. The list with junitFiles is populated with a SplFileObject for every process/thread it will be running on, but none of the temp files are actually generated/written.